Added 3 rooms, a key, a thing, a door, and some description.
I added a gold key which will unlock the bathroom door.  It is stuck in
the toilet.  I also added the east hallway, a living room, and a lawn.
The thing I added is a toilet in the bathroom which is where the key is
stuck.  I added the front door in the living room leading out to the
lawn.