Added 2 things, finished first puzzle.
I added a sink and a bathtub, just wanted to add a couple more things
in the bathroom so it wouldn’t be so plain.  And I figured out how to
make it so that the player would have to unclog the toilet by flushing
it and the key would come up.  Then you would’ve finished the first
puzzle.