Finished the second puzzle, and added more windows.
I added the windows that are needed to be able to solve the second
puzzle.  I added the mailbox and the paper inside it.  It is a clue for
where to go.  I added the bedroom window (which is the ‘key’ to this
puzzle) and the living room window, and a window in the west hallway.