I added a key, and the way to unlock the hidden door.
I added a silver key to be able to unlock the hidden door in the
closet.  You have to wear the tuxedo then the hidden door will be
revealed behind where the tuxedo was supposed to be.  And that leads to
the tunnel which leads to the church to stop the marriage.