I added more rooms, the NPC, and a door.
I added more rooms on the bottom level making the underground tunnel a
maze.  The NPC is Victoria’s dog, as you go into each room, you have to
listen to the dog because you will be able to hear whether you’re
getting closer or farther from it.  All the rooms of the underground
tunnel are dark rooms.  When you finally get to the dog, he leads you
upstairs, and you will be in the back room of a church and a changing
room.  I added a wooden door between the changing room and the back
room.  It is unlocked but closed.